Possible NULL de-reference on fullsceen check
Also quiet some other minor warnings

+++ source/blender/editors/space_node/drawnode.c
@@ -390,7 +390,7 @@ static void node_draw_frame_label(bNodeTree *ntree, bNode *node, const float asp
 	float width, ascender;
 	float x, y;
 	const int font_size = data->label_size / aspect;
-	const float margin = NODE_DY / 4;
+	const float margin = (float)(NODE_DY / 4);
 
 	nodeLabel(ntree, node, label, sizeof(label));
